book_recommender/requirements.txt
flask>=3.0.0
flask_sqlalchemy>=3.1.0
pydantic>=2.0.0
pytest>=9.0.0
google-generativeai>=0.7.0
